Feature Complete: Multithreading
Added process manager for variable worker threads. Worker threads do not
lock the main thread and can be deleted from the queue both before and
during operation. Data/Operation mutexes are disabled when workers are
running (to stop potentially deleting an object that is in queue to be
operated on).