Added Operation Settings
Implemented settings for Operation UserScripts. Ranges and defaults can
be set by creating a dictionary of expected settings and their respected
SettingsObject (currently only IntegerSettingObject is implemented).
When running a UserScript a dialog is generated that contains all
expected settings. The edited (or default) values are then submitted to
the worker for operation. Also implemented DataOut object naming by
simply setting the DataObject.name parameter in the script before
appending to DataOut.